backend/requirements.txt
fastapi[all]
pydantic
python-dotenv
groq
pandas
numpy
joblib
matplotlib
seaborn
scikit-learn
imbalanced-learn
xgboost
tensorflow
uvicorn
